requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
django-celery-beat>=2.5,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
channels>=4.0,<5.0
daphne>=4.0,<5.0
whitenoise>=6.5,<7.0
